refactor(issue-539): separate basic type measure traits
