+++ model/geo_contracts/src/geometry/core/mod.rs
@@ -93,8 +93,8 @@ pub use nurbs_surface_3d_traits::{
 };
 pub use plane3d_traits::Plane3DProperties;
 pub use point_traits::{
-    Point2DConstructor, Point2DCore, Point2DMeasure, Point2DProperties, Point3DConstructor,
-    Point3DCore, Point3DMeasure, Point3DProperties,
+    Point2DConstructor, Point2DCore, Point2DProperties, Point3DConstructor, Point3DCore,
+    Point3DProperties,
 };
 pub use ray_traits::{
     Ray2DConstructor, Ray2DCore, Ray2DMeasure, Ray2DProperties, Ray3DConstructor, Ray3DCore,
@@ -123,6 +123,6 @@ pub use triangle_traits::{
     Triangle3DConstructor, Triangle3DCore, Triangle3DMeasure, Triangle3DProperties,
 };
 pub use vector_traits::{
-    Vector2DConstructor, Vector2DCore, Vector2DMeasure, Vector2DProperties, Vector3DConstructor,
-    Vector3DCore, Vector3DMeasure, Vector3DProperties,
+    Vector2DConstructor, Vector2DCore, Vector2DProperties, Vector3DConstructor, Vector3DCore,
+    Vector3DProperties,
 };
